Shared MySQL charm library (#11)
* library initial import

* Add lib path

* strip vm specific code

* lint fix

* Adjust charm code to library usage

* avoid clash with concrete class name

* Add lib to test path

* adapted helper to library usage

* remove shared code
* add abstract methods implementation
* kept vm specific methods
* wait_until_mysql_connection non private method (used by charm code)
* clis calls raise custom exception

* fixed unit tests

* pleasing the linter

* refactor tests

* reorganize to mirror library usage
* add a bit more coverage

* opportunistic add actions from k8s

* fixed calls

* using `label` allow to display full topology

* fix docstring

* Fix action description

* fix lint

* improves library docstring
